로그인 로그아웃, 세션까지 끝 page 392

+++ src/main/java/com/example/demo/controller/BoardController.java
@@ -6,12 +6,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import com.example.demo.domain.Board;
+import com.example.demo.domain.Member;
 import com.example.demo.service.BoardService;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 
@@ -34,8 +38,17 @@ public class BoardController {
 //		model.addAttribute("boardList", boardList);
 //		return "getBoardList";
 //	}
+	
+	@ModelAttribute("member")
+	public Member setMember() {
+		return new Member();
+	}
+	
 	@RequestMapping("/getBoardList")
-	public String getBoardList(Model model, Board board) {
+	public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) {
+		if(member.getId() == null) {
+			return "redirect:login";
+		}
 		List<Board> boardList = boardService.getBoardList(board);
 		
 		model.addAttribute("boardList", boardList);
@@ -43,37 +56,53 @@ public class BoardController {
 	}
 	
 	@GetMapping("/insertBoard")
-	public String insertBoardView() {
+	public String insertBoardView(@ModelAttribute("member")Member member) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		return "insertBoard";
 	}
 	
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member")Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.insertBoard(board);
 		return "redirect:/getBoardList";
 	}
 	
 	@GetMapping("/getBoard")
-	public String getBoard(Board board, Model model) {
+	public String getBoard(@ModelAttribute("member")Member member, Board board, Model model) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		model.addAttribute("board", boardService.getBoard(board));
 		return "getBoard";
 	}
 	
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member")Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.updateBoard(board);
 		return "forward:/getBoardList";
 		//there is no difference between / and not used /
 	}
 	
 	@RequestMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member")Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
 	
 	@GetMapping("/hello")
 	public void hello(Model model) {
+		
 		model.addAttribute("greeting", "Hello, im using my first thymeleaf");
 		
 	}
